Ivan folder: clicking error message now highlights the error columns in Monaco editor
